[refactor] Quick split of our components into a few subfolders
Also add a link to the homepage on the footer's textualize logo
