+++ src/components/logo.tsx
@@ -1,3 +1,5 @@
+import React from "react"
+
 interface Props {
     isFooter?: boolean
 }
@@ -6,7 +8,7 @@ export const Logo = ({ isFooter }: Props) => (
         <svg className={`logo logo__white ${isFooter ? "logo--footer" : ""}`}>
             <use xlinkHref="#icon-logo-white" />
         </svg>
-        {!isFooter && (
+        {isFooter ? null : (
             <svg className={`logo logo__black`}>
                 <use xlinkHref="#icon-logo-black" />
             </svg>
