+++ src/pages/[projectId]/gallery/[[...gallerySegments]].tsx
@@ -2,7 +2,7 @@ import React from "react"
 import type { GetStaticPaths, GetStaticProps } from "next"
 import Head from "next/head"
 import { GetStaticPropsResult } from "next/types"
-import { GalleryIndex } from "../../../components/gallery-index"
+import { GalleryIndex } from "../../../components/page-specific/gallery/gallery-index"
 import * as githubBackendServices from "../../../services/backend/github"
 import * as galleryProjectsBackendServices from "../../../services/backend/projects-galleries"
 import { ProjectGalleryPageProps } from "../../../services/backend/projects-galleries"
