+++ src/pages/_app.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import type { AppProps } from "next/app"
 import Head from "next/head"
 import { useRouter } from "next/router"
-import { Layout } from "../components/layout"
+import { Layout } from "../components/layout/layout"
 import { initAnalytics } from "../services/frontend/analytics"
 import * as themeServices from "../services/shared/theme"
 import "../style/index.scss"
