+++ src/pages/index.tsx
@@ -1,8 +1,8 @@
 import React from "react"
 import type { GetStaticProps } from "next"
-import { Hero } from "../components/hero"
 import { MailList } from "../components/mail-list"
-import { ProjectsIndex } from "../components/projects-index"
+import { Hero } from "../components/page-specific/home/hero"
+import { ProjectsIndex } from "../components/page-specific/home/projects-index"
 import type { ProjectData } from "../domain"
 import * as githubBackendServices from "../services/backend/github"
 import * as projectsBackendServices from "../services/backend/projects"
